+++ Web/App_Start/FilterConfig.cs
@@ -7,8 +7,6 @@
         public static void RegisterGlobalFilters(GlobalFilterCollection filters)
         {
             filters.Add(new HandleErrorAttribute());
-            // TODO This doesn't do anything?
-            filters.Add(new RequireHttpsAttribute());
         }
     }
 }